Refactor big file into smaller files
