+++ frontend/lib/ui/dashboard/dashboard_body.dart
@@ -1,15 +1,9 @@
-import 'package:dartx/dartx.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:go_router/go_router.dart';
-import 'package:intersperse/intersperse.dart';
-import 'package:yaru_widgets/widgets.dart';
 
-import '../../models/artefact.dart';
 import '../../models/stage_name.dart';
-import '../../providers/stage_artefacts.dart';
 import '../../routing.dart';
 import '../spacing.dart';
+import 'stage_column.dart';
 
 class DashboardBody extends StatelessWidget {
   const DashboardBody({Key? key}) : super(key: key);
@@ -22,94 +16,9 @@ class DashboardBody extends StatelessWidget {
         horizontal: Spacing.pageHorizontalPadding,
       ),
       scrollDirection: Axis.horizontal,
-      itemBuilder: (_, i) => _StageColumn(stage: stages[i]),
+      itemBuilder: (_, i) => StageColumn(stage: stages[i]),
       separatorBuilder: (_, __) => const SizedBox(width: Spacing.level5),
       itemCount: stages.length,
-    );
-  }
-}
-
-class _StageColumn extends ConsumerWidget {
-  const _StageColumn({Key? key, required this.stage}) : super(key: key);
-
-  final StageName stage;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final family = AppRoutes.familyFromContext(context);
-    final artefacts = ref.watch(stageArtefactsProvider(family, stage));
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          stage.name.capitalize(),
-          style: Theme.of(context).textTheme.headlineSmall,
-        ),
-        const SizedBox(height: Spacing.level4),
-        Expanded(
-          child: SizedBox(
-            width: _ArtefactCard.width,
-            child: artefacts.when(
-              data: (artefacts) => ListView.separated(
-                itemBuilder: (_, i) => _ArtefactCard(artefact: artefacts[i]),
-                separatorBuilder: (_, __) =>
-                    const SizedBox(height: Spacing.level4),
-                itemCount: artefacts.length,
-              ),
-              loading: () =>
-                  const Center(child: YaruCircularProgressIndicator()),
-              error: (error, stackTrace) =>
-                  Text('Failed to fetch artefacts: $error'),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _ArtefactCard extends ConsumerWidget {
-  const _ArtefactCard({Key? key, required this.artefact}) : super(key: key);
-
-  final Artefact artefact;
-  static const double width = 320;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return GestureDetector(
-      onTap: () {
-        final currentRoute = GoRouterState.of(context).fullPath;
-        context.go('$currentRoute/${artefact.id}');
-      },
-      child: Card(
-        margin: const EdgeInsets.all(0),
-        elevation: 0,
-        shape: RoundedRectangleBorder(
-          side: BorderSide(color: Theme.of(context).colorScheme.outline),
-          borderRadius: BorderRadius.circular(2.25),
-        ),
-        child: Container(
-          width: width,
-          height: 156,
-          padding: const EdgeInsets.all(Spacing.level4),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                artefact.name,
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-              const SizedBox(height: Spacing.level2),
-              ...artefact.details.entries
-                  .map<Widget>(
-                    (detail) => Text('${detail.key}: ${detail.value}'),
-                  )
-                  .intersperse(const SizedBox(height: Spacing.level2)),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
